+++ src/components/button/Button.ts
@@ -1,10 +1,10 @@
 import Handlebars from "handlebars";
-import "./button.pcss";
+import "./Button.pcss";
 
-const buttonTemplate = `<button id="{{id}}" class="button"
-  {{#if disabled}} disabled {{/if}}
->
-  {{text}}
-</button>`;
+const buttonTemplate = `
+  <button id="{{id}}" class="button" {{#if disabled}} disabled {{/if}}>
+    {{text}}
+  </button>
+`;
 
 export const Button = Handlebars.compile(buttonTemplate);
