+++ src/pages/messenger/components/chat-list/search-bar/SearchBar.pcss
