+++ src/pages/user-profile/UserProfile.ts
@@ -15,12 +15,12 @@ interface IProps {
 
 export const UserProfilePage = ({ user }: IProps) => {
   const profileData = [
-    { label: "Почта", value: "pochta@yandex.ru" },
-    { label: "Логин", value: "ivanivanov" },
-    { label: "Имя", value: "Иван" },
-    { label: "Фамилия", value: "Иванов" },
-    { label: "Имя в чате", value: "Иван" },
-    { label: "Телефон", value: "+7 (909) 967 30 30" },
+    { label: "Почта", value: user.email },
+    { label: "Логин", value: user.login },
+    { label: "Имя", value: user.firstName },
+    { label: "Фамилия", value: user.lastName },
+    { label: "Имя в чате", value: user.chatName },
+    { label: "Телефон", value: user.phone },
   ];
 
   const profileList = profileData.map((item) => ProfileListItem(item)).join("");
